Harden core edge cases, deduplicate validation, and extend regression tests
Fixes verified by the new tests:
- StageClock widened long frames to float before multiplying, losing
  single-frame precision past 2^24 frames; the double product now stays
  integral-exact before narrowing to the float32 tick grid.
- A huge stun/displace duration rounded to Long.MAX_VALUE frames and
  wrapped the expiry frame negative; it now clamps within the local
  frame counter's reachable range.
- GridMap had no dimension bound, so width * height could overflow int
  into NegativeArraySizeException; dimensions are capped at 512, matching
  the scenario codec's import limit.
- The collision DDA swept axes with denormal displacements whose
  crossing times could tie at Infinity and trigger the three-cell corner
  rule for a pure-axis move; near-zero components are now stationary.

Deduplication:
- GlobalFrameSequencer shares the non-negative/consecutive frame guard
  between Stage and PathfindingSimulator.
- Vec2f.isFinite replaces three private copies in Route, UnitConfig,
  and Checkpoint.
- NextNodeSmoother's diagonal and orthogonal passes merge into one
  parameterized loop.
- The stable-node radius is a named constant, and visit-history writes
  go through UnitState.markVisitedTileCenter.

Tests grow from 49 to 58: avoidance influence band (cardinal, diagonal
averaging, projection, below-margin zero), tile-center visit steering,
multi/terminal portal route distances, post-portal completion,
injection rejection for vanished/completed units, play-time rebasing,
capacity-one trace ring, stun clamping, and map dimension bounds.

+++ src/main/StageClock.java
@@ -53,6 +53,10 @@ public final class StageClock {
     }
 
     private static float elapsedSeconds(long frames) {
-        return frames * F32.DT;
+        // Widen before multiplying: frames * F32.DT would promote the long to
+        // float first, silently quantizing every frame count past 2^24. The
+        // double product stays integral-exact for any feasible frame count and
+        // narrows to the same float32 tick grid deterministically.
+        return (float) frames * F32.DT;
     }
 }
